[BREAKING] renaming `intl:main` to `service:intl`
